+++ src/movie.ts
@@ -432,8 +432,8 @@ export class Movie {
       return
     }
 
-    window.requestAnimationFrame(timestamp => {
-      this._render(repeat, timestamp)
+    window.requestAnimationFrame(() => {
+      this._render(repeat, undefined, done)
     }) // TODO: research performance cost
   }