+++ src/movie.ts
@@ -2,7 +2,6 @@
  * @module movie
  */
 
-import { AudioContext } from 'standardized-audio-context'
 import { subscribe, publish } from './event'
 import { Dynamic, val, clearCachedValues, applyOptions, watchPublic } from './util'
 import { Base as BaseLayer, Audio as AudioLayer, Video as VideoLayer, Visual } from './layer/index' // `Media` mixins