+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.8.4] - 2022-02-23
 ### Fixed
 - **Major memory leak.**
 - `Movie#play()` not resolving.
@@ -201,6 +201,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Gaussian blur
   - Transform
 
+[0.8.4]: https://github.com/clabe45/vidar/compare/v0.8.3...v0.8.4
 [0.8.3]: https://github.com/clabe45/vidar/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/clabe45/vidar/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/clabe45/vidar/compare/v0.8.0...v0.8.1